+++ cmd/ttn-lw-cli/commands/root.go
@@ -21,6 +21,7 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"fmt"
+	"math/rand"
 	"net/http"
 	"net/url"
 	"os"
@@ -181,6 +182,15 @@ func preRun(tasks ...func() error) func(cmd *cobra.Command, args []string) error
 		if config.DumpRequests {
 			api.SetDumpRequests(true)
 		}
+
+		// initializes seed for random related operations
+		rand.Seed(time.Now().UnixNano())
+
+		api.SetRetryMax(config.Retry.Max)
+		api.SetRetryDefaultTimeout(config.Retry.DefaultTimeout)
+		api.SetRetryEnableMetadata(config.Retry.EnableMetatada)
+		api.SetRetryJitter(config.Retry.Jitter)
+
 		if config.CA != "" {
 			pemBytes, err := os.ReadFile(config.CA)
 			if err != nil {