+++ cmd/ttn-lw-cli/internal/api/grpc.go
@@ -26,16 +26,21 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/rpcclient"
 	"go.thethings.network/lorawan-stack/v3/pkg/rpcmetadata"
 	"go.thethings.network/lorawan-stack/v3/pkg/rpcmiddleware/rpclog"
+	"go.thethings.network/lorawan-stack/v3/pkg/rpcmiddleware/rpcretry"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/connectivity"
 	"google.golang.org/grpc/credentials"
 )
 
 var (
-	withInsecure bool
-	tlsConfig    *tls.Config
-	auth         *rpcmetadata.MD
-	withDump     bool
+	withInsecure        bool
+	tlsConfig           *tls.Config
+	auth                *rpcmetadata.MD
+	withDump            bool
+	retryMax            uint
+	retryDefaultTimeout time.Duration
+	retryEnableMetadata bool
+	retryJitter         float64
 )
 
 // SetLogger sets the default API logger
@@ -51,6 +56,26 @@ func SetInsecure(insecure bool) {
 // SetDumpRequests configures the API options to dump gRPC requests.
 func SetDumpRequests(dump bool) {
 	withDump = dump
+}
+
+// SetRetryMax configures the amount of time the client will retry the request.
+func SetRetryMax(rm uint) {
+	retryMax = rm
+}
+
+// SetRetryDefaultTimeout configures the default timeout before making a retry in a failed request.
+func SetRetryDefaultTimeout(t time.Duration) {
+	retryDefaultTimeout = t
+}
+
+// SetRetryEnableMetadata configures if the retry procedure will read the request's metadata or not.
+func SetRetryEnableMetadata(b bool) {
+	retryEnableMetadata = b
+}
+
+// SetRetryJitter configures the fraction to be used in the deviation procedure of the rpcretry timeout.
+func SetRetryJitter(f float64) {
+	retryJitter = f
 }
 
 // AddCA adds the CA certificate file.
@@ -107,6 +132,13 @@ func GetDialOptions() (opts []grpc.DialOption) {
 	if withDump {
 		opts = append(opts, grpc.WithChainUnaryInterceptor(requestInterceptor))
 	}
+
+	opts = append(opts, grpc.WithChainUnaryInterceptor(rpcretry.UnaryClientInterceptor(
+		rpcretry.WithMax(retryMax),
+		rpcretry.WithDefaultTimeout(retryDefaultTimeout),
+		rpcretry.UseMetadata(retryEnableMetadata),
+		rpcretry.WithJitter(retryJitter),
+	)))
 	return
 }